delta_ingestion_pipeline_fixed/requirements.txt
pyspark
delta-spark
faker
pandas
apscheduler
pytz
yagmail